requirements.txt
aiofiles==22.1.0
aiosmtplib==1.1.7
anyio==3.6.2
bcrypt==4.0.1
blinker==1.5
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cloudinary==1.30.0
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.88.0
fastapi-login==1.8.3
fastapi-mail==1.2.2
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
orjson==3.8.3
passlib==1.7.4
psycopg2-binary==2.9.5
pydantic==1.10.2
PyJWT==2.6.0
python-decouple==3.6
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.7
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqladmin==0.8.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a31
sqlalchemy_database==0.1.0
sqlmodel==0.0.8
starlette==0.21.0
twilio==7.16.0
typing_extensions==4.4.0
ujson==5.6.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
WTForms==3.0.1
